dotfiles_processor::context
Struct Context

pub struct Context { /* private fields */ }

A context represents an environment in which defaults can be overriden, it can be thought of as the context of an individual configuration file to apply.

Notice that contexts can be built on top of one another, so that defaults can be overriden multiple times, and thus have some sort of configuration inheritance.

Implementations

impl Context

pub fn defaults(&self) -> &HashMap<String, Settings>

The default overrides for the current context.

pub fn actions(&self) -> &Vec<KnownAction<'static>>

The list of actions parsed from this file.

pub fn file(&self) -> &PathBuf

The absolute path to the file to which this context corresponds.

impl Context

pub fn subcontext(&self, file: &Path) -> Result<Context, DotfilesError>

pub fn get_default(&self, dir: &str, setting: &str) -> Option<&Setting>

pub fn parse_file(&mut self) -> Result<(), DotfilesError>

pub fn run_actions(context: Context) -> Result<(), DotfilesError>

Runs the actions in this context and consumes the context.